Move sort key desc to argument
